Development ss16

+++ latex/day4/problems/listings/functional_map_sig.cpp
@@ -1,2 +1,2 @@
 template<typename SeqT>
-void map(SeqT &seq, int(*func)(int i));
+void map(SeqT &seq, double(*func)(double i));
