+++ latex/day4/problems/listings/functional_reduce_sig.cpp
@@ -1,2 +1,2 @@
 template<typename SeqT, typename RetT>
-RetT reduce(SeqT &seq, RetT initialVal, RetT(*func)(RetT &i, int j));
+RetT reduce(SeqT &seq, RetT initialVal, RetT(*func)(RetT &i, double j));
